Inline methods to avoid capturing blocks
When you capture the block, you have to declare what is going to be passed into it and if it will return anything.
By not capturing it, it can be more "dynamic".
I will probably pull some of it into a macro to avoid all the duplication but I would like to keep it verbose for now.