Add json plugin
This is different than Roda's. Roda's plugin handle the result of the block and matches it against the registered types. I'm not confident that is easily do-able in Crystal and I'm also not sure that is even the best API. I think it's more obvious and fitting in crystal to have an explicit `json` method call with the thing you expect to be turned into json. That way we can just call the generic `to_json` method and let the user deal with it if it's not what they want.