NamedRoutes plugin
Figuring out how to connect the string to a block of code was tough.
I find previous_def to be confusing about when it works.
I had hoped that an empty `handle_named_route` method would work
but it didn't.

I think the difficulty with all of this has been trying to make sure
that the blocks are all ran in the instance level of the app.
If I just captured the blocks and stored them then it wouldn't work.

I found that I need the `route` block to return something
and it get rendered.
So that's a TODO